Delete Fresh framework app
