+++ deno.json
@@ -9,17 +9,11 @@
   "imports": {
     "@/": "./",
     "@std/http": "jsr:@std/http@^1.0.13",
-    "@fresh/core": "jsr:@fresh/core@2.0.0-alpha.29",
     "@std/dotenv": "jsr:@std/dotenv@^0.225.3",
     "preact": "npm:preact@10.25.4",
     "@preact/signals": "npm:@preact/signals@2.0.1",
     "@preact/signals-core": "npm:@preact/signals-core@1.8.0",
-    "@twind/core": "npm:@twind/core@1.1.3",
-    "@twind/preset-tailwind": "npm:@twind/preset-tailwind@1.1.4",
-    "@twind/preset-autoprefix": "npm:@twind/preset-autoprefix@1.0.7",
-    "@twind/aspect-ratio": "npm:@twind/aspect-ratio@0.1.4",
     "preact-helmet": "npm:preact-helmet@^4.0.0-alpha-3",
-    "preact-render-to-string": "npm:preact-render-to-string@^6.5.13",
-    "swr": "npm:swr@2.3.2"
+    "preact-render-to-string": "npm:preact-render-to-string@^6.5.13"
   }
 }
